requirements.txt
Flask==2.3.3
pdfplumber==0.9.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.3
pdfminer.six==20221105
Pillow==10.0.1
charset-normalizer==3.3.0
cryptography==41.0.7
reportlab==4.0.4
pandas==2.0.3
openpyxl==3.1.2
gunicorn==21.2.0
psycopg2-binary==2.9.7
PyMySQL==1.1.0
flask-limiter==3.5.0
flask-talisman==1.1.0
requests==2.31.0
numpy==1.24.3
chardet==5.2.0
colorama==0.4.6
et-xmlfile==1.1.0
pypdf2==3.0.1